fix: move EnsureOtelCollectorConfig to project package
Address review feedback: move otel collector config generation to
internal/cli/agent/project/project.go alongside other Regenerate*
functions, with matching signature. Move test to project_test.go.
